Implement distinct
Removed the validation code for multiple distinct calls. I'm not
entirely sure it's necessary to begin with. I'm not sure of the
consequences of doing this, but the reason is because it would be too
hard to figure out how to use distinct with math or support nested
functions in the future without relaxing the conditions on distinct a
bit.

As long as we don't end up in a deadlock at any point, I don't think it
matters if the query doesn't make sense to me. It might be useful to
somebody.