Implemented symbol resolution!
Just need to replace the IteratorCreator implementation with creating
the individual shards and fix up link auxiliary fields so some of the
errors get thrown at compilation time instead of linking time.

But, this is huge progress! Almost everything from coordinator is now in
query and it mostly conforms to specifications.